IE compat for change function
IE7,8 has an issue with the change function.  Although it fires the
change event when the radio button changes, the $('input:checked')
selector later in the code is still the old value.

Online, a common fix is to apply a click listener and do a blur/focus
... which didn't work for me.  However, it appears that setting the
checked property fixes things.