Fix isColorSection variable
Our desired behaviour is that we tell the gallery widget to refresh the
product image only when we're refreshing in the context of a colour
(i.e. no need to change image if user is changing sizes).

However, because $sections.each both sets isColorSection and triggers
refresh.sku.xxx, we basically always trigger both refresh.sku.color and
refresh.sku.size events, which gives us incorrect behaviour.

Move isColorSection variable outside of the $sections.each
(necessitating clients to pass in $this) to prevent triggering the
incorrect event.